4th upload
The deer, grass, and trees all reproduce. I think the plants are going into an ArrayList, but it's hard to tell on the simulator, because I can't check individual locations. When I create a field and a location and add a plant to it, it goes to the ArrayList, though. Also, I don't know if the deer and the plants are coexisting in the same square, because only one color comes through. That needs to be tested. The lower probability rates of creation for all the organisms allows the simulation to run for more steps before it has a memory error...what does that mean??